Moved parser functionaity into page class
Created page class that manages and stores page data. Each method in the
page class returns a specific piece of data from the page. If that data
has not yet been parsed the method parses it and returns it, otherwise
it returns the previously parsed data. Also combined weight and
dimensions into one function, because the process to retrieve and parse
each is nearly identical so we can preform them at the same time.